+++ bangfuse/metadata_utils.go
@@ -28,7 +28,7 @@ func MetadataToFuseAttr(inum uint64, meta *pb.InodeMeta, out *fuse.Attr) {
 	out.Mtimensec = uint32(meta.MtimeNs % 1e9)
 	out.Ctime = uint64(meta.CtimeNs / 1e9)
 	out.Ctimensec = uint32(meta.CtimeNs % 1e9)
-	out.Blksize = GetChunkSize() // I/O hint for large reads
+	out.Blksize = uint32(GetChunkSize()) // I/O hint for large reads
 
 	// st_blocks must be in 512-byte units (POSIX)
 	out.Blocks = (meta.Size + 511) / 512
